Release 1.12.0 (#14)
* refactor: migrate cart routes and UI to use cart-items endpoints

* refactor and add field to customer form

* add comprehensive installation guide for local environments

* refactor: update environment variables and improve HTTP request syntax in user registration

* feat: add DNI field to customer forms and validators; replace create_user page with ask page

* refactor: add AdminLayout component and update imports across customer, message, order, and showcase pages

* test: add unit tests for AdminLayout component

* chore: update eslint config

* feat: implement AdminSideMenu component with dynamic menu options and routing

* refactor: update message storage structure and enhance delete confirmation dialog

* feat: add Spanish language support for validation messages and update message structure

* feat: implement stock management features including create, store, and image upload functionality

* feat: enhance ImageUpload component with delete functionality and confirmation dialog tests

* feat: add provider selection to stock creation and update cart item view route

* remove unused import

+++ database/migrations/1751317954868_add_factory_code_to_products_table.ts
@@ -11,9 +11,11 @@ export default class extends BaseSchema {
   }
 
   async down() {
-    this.schema.alterTable(this.tableName, (table) => {
-      table.dropColumn('factory_code')
-      table.dropIndex('factory_code')
-    })
+    if (await this.schema.hasColumn(this.tableName, 'factory_code')) {
+      this.schema.alterTable(this.tableName, (table) => {
+        table.dropIndex('factory_code')
+        table.dropColumn('factory_code')
+      })
+    }
   }
 }
